+++ engine/engine/src/main/groovy/com/virtusa/gto/nyql/engine/pool/QJdbcPool.groovy
@@ -1,12 +1,15 @@
 package com.virtusa.gto.nyql.engine.pool
 
+import com.virtusa.gto.nyql.configs.Configurations
 import com.virtusa.gto.nyql.exceptions.NyException
 /**
  * @author IWEERARATHNA
  */
 interface QJdbcPool extends QJdbcPoolFetcher {
 
-    void init(Map options) throws NyException
+    String getName()
+
+    void init(Map options, Configurations configurations) throws NyException
 
     void shutdown() throws NyException