+++ engine/engine/src/main/groovy/com/virtusa/gto/nyql/engine/repo/Caching.groovy
@@ -53,6 +53,7 @@ class Caching implements Closeable {
                     }
                 }
                 LOGGER.info('Compilation successful!')
+                LOGGER.info('-'*80)
             }
         }
     }